had to add my own adapter for HistoricalData

+++ app/src/main/java/cryptomarket/eoinahern/ie/cryptomarket/data/models/TimeInstanceData.kt
@@ -4,11 +4,11 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class TimeInstanceData(
-
 		val time: Long,
 		val close: Float,
 		val high: Float,
 		val low: Float,
 		val open: Float,
 		val volumefrom: Float,
-		val volumeto: Float)
+		val volumeto: Double
+)
